requirements.txt
python-dotenv
streamlit
langchain
langchain_google_genai
google-generativeai
pandas
scikit-learn